Major refactoring to improve CreatureAccessoryLootBox.
CreatureAccessoryLootBox is now an ERC1155 token.
The CreatureAccessory unboxing logic is now a library as that was
the only way to fit everything into the contract.
It does not implement IFactory due to API clashes and it does not
implement ERC1155Tradable due to needing to assign its own token IDs
for LootBox tokens.
It does properly mint and unbox/burn LootBox tokens.
I'm not sure about the API though, we will need to discuss that.
And the migrations need fixing.